src/api/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pydantic==2.8.2
asyncpg==0.29.0
pygdal==3.2.2.*
simplekml==1.3.6
httpx
gunicorn
PyJWT
